sample GL project using static library

+++ rc3dksamplegl/RC3DKSampleGL/CalibrationStep1.m
@@ -82,7 +82,7 @@
 
 - (void) sensorFusionError:(NSError*)error
 {
-    NSLog(@"SENSOR FUSION ERROR %i", error.code);
+    NSLog(@"SENSOR FUSION ERROR %li", (long)error.code);
 }
 
 - (void) calibrationFinished
